QuantAnalyticsDashboard/requirements.txt
streamlit
pandas
numpy
plotly
statsmodels
websockets
aiohttp
pyyaml
scipy
kaleido
watchdog
